Update and move statup class into FW dll.
