+++ Development/AzRAdmin/Presentations/AzR.Web/App_Start/WebApiConfig.cs
@@ -1,8 +1,8 @@
-﻿using AzR.Web.Handlers;
-using Microsoft.Owin.Security.OAuth;
+﻿using Microsoft.Owin.Security.OAuth;
 using System.Net.Http.Headers;
 using System.Web.Http;
 using System.Web.Http.Cors;
+using AzR.WebFw.Handlers;
 
 namespace AzR.Web
 {
