+++ Development/AzRAdmin/Presentations/AzR.Web/Areas/Api/Controllers/V1/UserPrivilegeV1Controller.cs
@@ -2,11 +2,11 @@
 using AzR.Core.ViewModels.Admin;
 using AzR.Utilities.Exentions;
 using AzR.Web.Controllers;
-using AzR.Web.Filters;
 using System;
 using System.Threading.Tasks;
 using System.Web.Http;
 using AzR.WebFw.Controllers;
+using AzR.WebFw.Filters;
 
 namespace AzR.Web.Areas.Api.Controllers.V1
 {
