+++ Development/AzRAdmin/Presentations/AzR.Web/ErrorLog.txt
@@ -19,3 +19,4 @@
 20/May/2018 11:26:: PcUniqueNumber.cs:: get_GetUserAgentInfo::157 :: Object reference not set to an instance of an object.
 20/May/2018 13:56:: PcUniqueNumber.cs:: get_GetUserAgentInfo::157 :: Object reference not set to an instance of an object.
 20/May/2018 13:56:: PcUniqueNumber.cs:: get_GetUserAgentInfo::157 :: Object reference not set to an instance of an object.
+22/May/2018 14:37:: PcUniqueNumber.cs:: get_GetUserAgentInfo::157 :: Object reference not set to an instance of an object.
