--- Development/AzRAdmin/Presentations/AzR.Web/Providers/GenericIdentityExtensions.cs
@@ -1,63 +1,0 @@
-﻿using System.Security.Claims;
-using System.Security.Principal;
-
-namespace AzR.Web.Providers
-{
-    /*
-     identity.AddClaim(new Claim("oauth:client", client));
-
-    var currentUser = RequestContext.Principal;
-    var client = currentUser.Identity.GetClientId();
-     *
-     */
-
-    public static class GenericIdentityExtensions
-    {
-        const string ClientIdentifier = "oauth:client";
-
-        /// <summary>
-        /// Set the client id claim
-        /// </summary>
-        /// <param name="identity"></param>
-        /// <returns></returns>
-        public static bool SetClientId(this IIdentity identity, string clientId)
-        {
-            if (identity != null)
-            {
-                var claimsIdentity = identity as ClaimsIdentity;
-                if (claimsIdentity != null)
-                {
-                    claimsIdentity.AddClaim(new Claim(ClientIdentifier, clientId));
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Return the client id claim
-        /// </summary>
-        /// <param name="identity"></param>
-        /// <returns></returns>
-        public static string GetClientId(this IIdentity identity)
-        {
-            if (identity != null)
-            {
-                var claimsIdentity = identity as ClaimsIdentity;
-                if (claimsIdentity != null)
-                {
-                    return claimsIdentity.FindFirstOrEmpty(ClientIdentifier);
-                }
-            }
-            return string.Empty;
-        }
-        /// <summary>
-        /// Retrieves the first claim that is matched by the specified type if it exists, String.Empty otherwise.
-        /// </summary>
-        public static string FindFirstOrEmpty(this ClaimsIdentity identity, string claimType)
-        {
-            var claim = identity.FindFirst(claimType);
-            return claim == null ? string.Empty : claim.Value;
-        }
-    }
-}
