+++ Development/AzRAdmin/Presentations/AzR.WebFw/Handlers/ContextItemHandler.cs
@@ -3,7 +3,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using System.Web;
 
-namespace AzR.Web.Handlers
+namespace AzR.WebFw.Handlers
 {
     public class ContextItemHandler : DelegatingHandler
     {
