+++ Development/AzRAdmin/Presentations/AzR.WebFw/packages.config
@@ -4,12 +4,22 @@
   <package id="AutoMapper" version="6.2.2" targetFramework="net452" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net452" />
   <package id="EntityFramework6.Npgsql" version="3.1.1" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Identity.Owin" version="2.2.1" targetFramework="net452" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.6" targetFramework="net452" />
   <package id="Microsoft.AspNet.Razor" version="3.2.6" targetFramework="net452" />
+  <package id="Microsoft.AspNet.SignalR.Core" version="2.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.Web.Optimization" version="1.1.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net452" />
+  <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.6" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.6" targetFramework="net452" />
+  <package id="Microsoft.Owin" version="4.0.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Host.SystemWeb" version="4.0.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security" version="4.0.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security.Cookies" version="4.0.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security.OAuth" version="4.0.0" targetFramework="net452" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
   <package id="MySql.Data" version="6.9.11" targetFramework="net452" />
   <package id="MySql.Data.Entity" version="6.9.11" targetFramework="net452" />
@@ -17,6 +27,7 @@
   <package id="Npgsql" version="3.2.7" targetFramework="net452" />
   <package id="Oracle.ManagedDataAccess" version="12.2.1100" targetFramework="net452" />
   <package id="Oracle.ManagedDataAccess.EntityFramework" version="12.2.1100" targetFramework="net452" />
+  <package id="Owin" version="1.0" targetFramework="net452" />
   <package id="PagedList" version="1.17.0.0" targetFramework="net452" />
   <package id="PagedList.Mvc" version="4.5.0.0" targetFramework="net452" />
   <package id="System.Data.SQLite" version="1.0.108.0" targetFramework="net452" />
